Module chunkwm::api

The api module contains the raw chunkwm_api struct, and contains a CVar struct to create a handle to a specific CVar.

Structs

API
The API struct is a raw type (i.e. the same as used in C code), and contains methods for creating, finding, and getting CVars.

CVar
The CVar struct allows you to easily create a handle to CVars, that might be updated. You can use all types that implement Display and FromStr as value type for the CVar. You can therefore easily implement your own CVar compatible types.

NumericBool
The NumericBool is often used in a chunkwm config. It is a boolean type that is either 0 or 1 (false or true respectively). This struct allows you to create a CVar that is a numeric bool, and easily get its value.

Enums

LogLevel
The LogLevel defines how something should be logged for chunkwm. See the chunkwm readme.