Struct chunkwm::prelude::CVar

pub struct CVar<T: FromStr + Display> {
    pub name: &'static str,
    // some fields omitted
}

The CVar struct allows you to easily create a handle to CVars, that might be updated. You can use all types that implement Display and FromStr as value type for the CVar. You can therefore easily implement your own CVar compatible types.

Example

pub struct Plugin {
    global_desktop_mode: CVar<String>,
    bsp_spawn_left: CVar<NumericBool>,
}

impl HandleEvent for Plugin {
    fn new(api: &'static API) -> Plugin {
        let global_desktop_mode = CVar::new("global_desktop_mode", &api).unwrap();
        let bsp_spawn_left = CVar::new("bsp_spawn_left", &api).unwrap();
        Plugin { global_desktop_mode, bsp_spawn_left }
    }

    fn handle(&mut self, _: Event) {
        self.global_desktop_mode.set_value(&"bsp".to_owned());
        assert_eq!(self.global_desktop_mode.get_value().unwrap(), "bsp".to_owned());

        self.bsp_spawn_left.set_value(&NumericBool::from(false));
        assert_eq!(self.bsp_spawn_left.get_value().unwrap().value, false);
    }
    // some methods omitted
}

Fields
name: &'static str

The CVar's name.

Methods

impl<T: FromStr + Display> CVar<T>

pub fn new(name: &'static str, api: &'static API) -> Result<Self, &'static str>

Create a new CVar handle without a value.

pub fn with_value(name: &'static str, value: &T, api: &'static API) -> Self

Create a new CVar handle with a value. The given value will be set.

pub fn set_value(&mut self, value: &T)

Set the value of the CVar.

pub fn get_value(&mut self) -> Result<T, <T as FromStr>::Err>

Get the value of the CVar.