Module chunkwm::raw

Structs

ChunkWMPlugin
The raw type ChunkWM uses for a plugin.

ChunkWMPluginDetails
The raw plugin_details type ChunkWM uses for getting information about a plugin.

ProcessSerialNumber
The process serial number. Both used as a raw type and as a Rust type.

RawApplication
The raw application representation. This is used in the C/C++ plugin, and shouldn't really be necessary for you to use in your Rust code.

RawDisplay
The raw Display.

RawObserver
The raw observer.

RawPayload
The raw payload representation. This is used in the C/C++ plugin, and shouldn't really be necessary for you to use in your Rust code.

RawSpace
The raw Space.

RawWindow
The raw window representation. This is used in the C/C++ plugin, and shouldn't really be necessary for you to use in your Rust code.

Type Definitions

AXError
An error type. This is not often used in Rust code.

AXUIElementRef
The reference to UI elements.

ApplicationRef
A reference to an RawApplication.

CGSSpaceID

CGSSpaceType

DisplayRef
A reference to a RawDisplay.

PID
The process identifier type.

PayloadRef
A reference to a RawPayload.

PluginBoolFunc
The function type used for the plugin's init function.

PluginFunc
The function type used for initialising and getting the plugin.

PluginMainFunc
The function type used for the plugin's main function.

PluginVoidFunc
The function type used for the plugin's deinit function.

SpaceRef
A reference to a RawSpace.

WindowRef
A reference to a RawWindow.